Moved FK check for it own file
